:art: [CHORE]entity 폴더명 domain으로 수정
